Move writeRiddlesToFile function to a separate file for consistency
