+++ riddels/CurdRiddels/create.js
@@ -1,16 +1,12 @@
 // Imports required modules for file system, input handling, paths, and styling
 import fs from "fs/promises";
 import readline from "readline-sync";
-import path from "path";
-import { fileURLToPath } from "url";
 import chalk from "chalk";
 import loadRiddleDatabase from "./read.js"
 import { getInputFromUser } from "../../systemOprtion/uiManager.js"
+import {writeRiddlesToFile} from "./saveRiddlesToDB.js"
 
-// Set up the path to the riddle database file
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-const dbPath = path.resolve(__dirname, "../riddelsDB/db.txt");
+
 
 
 
@@ -50,21 +46,7 @@ function createRiddle(riddles) {
 }
 
 
-/**
- * Save riddles array to the database file.
- * @param {Array} riddles - Array of all riddle objects to be saved.
- * @returns {Promise<void>}
- */
-export async function saveRiddlesToDB(riddles) {
-  try {
-    await fs.writeFile(dbPath, JSON.stringify(riddles, null, 2));
-    console.log(chalk.green("The new puzzle has been successfully entered into the database."))
-  }
-  catch (err) {
-    throw new Error("Failed to write to DB: " + err.message);
-  }
 
-}
 
 
 /**
@@ -77,10 +59,12 @@ export async function createMenager() {
     const riddles = await loadRiddleDatabase();
     const newObj = createRiddle(riddles);
     riddles.push(newObj);
-    await saveRiddlesToDB(riddles);
+    await writeRiddlesToFile(riddles);
   }
   catch (err) {
     throw new Error(chalk.red("Error: " + err.message));
   }
 
 }
+
+await createMenager()
